LayoutTests/media/modern-media-controls/macos-compact-inline-media-controls/macos-compact-inline-media-controls-buttons-styles.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>MacOSCompactInlineMediaControls</code> computed styles.");

window.jsTestIsAsync = true;

const mediaControls = new MacOSCompactInlineMediaControls({ width: 240, height: 300 });

const buttonPositions = [
    [mediaControls.playPauseButton, 6],
    [mediaControls.skipBackButton, 6],
    [mediaControls.muteButton, 6],
    [mediaControls.fullscreenButton, 6.5]
];

let style;
scheduler.frameDidFire = function()
{
    document.body.appendChild(mediaControls.element);

    if (buttonPositions.some(buttonPosition => buttonPosition[0].width == 0))
        return;

    for (let [button, y] of buttonPositions) {
        style = window.getComputedStyle(button.element);
        debug(button.constructor.name);
        shouldBeEqualToRGBAColor("style.backgroundColor", "rgba(255, 255, 255, 0.572)");
        shouldBeEqualToString("style.height", "25px");
        shouldBeEqualToString("style.webkitMaskPositionY", `${y}px`);
        debug("");
    }

    mediaControls.element.remove();
    finishMediaControlsTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
